Bug 699638(3): Fix param list key handling in pl_main_process_options()
param lists can be configured to handle keys in two ways: persistent keys, or
transient keys. For persistent keys, the param list contains a reference to the
key string, and does no management of the string memory.

Configured for transient key strings, the param list makes a copy of the string
and the string memory is managed along with the parameter list itself.

The two approaches cannot be mixed in the same param list.

The pl_main_process_options() code configures the param list to use transient
key strings (which is the default). But it was then copying the key strings,
before passing them to the param list API, and taking no action to manage
the string memory - causing memory leaks.

Since the param list (as configured) takes a copy of the key string, there's no
call for the calling code to do so, and not doing the local string copy
prevents leaking memory.