Rename all the PDF interpreter functions
Nancy noticed that the PDF interpreter function pdf_setlinewidth was
also defined in the pdfwrite device. Although this is a static function
it did rather point out the possibility of name clashes.

To prevent that, rename all the pdf_ interpreter functions to be pdfi_
instead.

I purposefully have not renamed the PDF object types, pdfwrite should
not be using the interpreter object types, and vice versa.