Separate colour space creation and setting
We need to create colour spaces for images, we don't want to duplicate
this (complex) code. So do the colour space creation and setting it in
the graphics state separately.

Extensively alter the error handling, we now always pass errors back up
until we reach the PDF oeprator, and decide then whether or not to stop
based on ctx->pdfstoponerror. Prevents confusion about whether a routine
needs to return an error or not.

Restore the saved offset when we get an error reading an ICC profile.

Add code to read the lookup table for an Indexed colour space from an
indirect reference. If the reference is to a string, not a stream, then
don't throw an error, even though this is technically illegal, just use
the string data.

Add the abbreviated colour spaces "G", "RGB, "CMYK" and "I" so that we
can create these appropriately.

Temporarily treat a /Separation space as DeviceGray. Allows files to
be processed.