Correct counting of uncompressed streams
When we need to know the size of an uncompressed stream, in order to
create a memory buffer, we read and decompress one byte at a time. But
that always returns a single byte, and the return value is not the
count of bytes read, just a success code.

So don't add the success code, just increment the counter. This corrects
a number of allocation errors.