Bug 699623: ICC profile creation file permissions check
The original fix missed the initial branch with an explicitly specified
directory.

This also uncovered problems with device profile reference counting, where
if profile creation failed we'd leave a pointer to the original profile
in place - either a dangling pointer, or leaving the ref count wrong.

To solve this, we NULL the relevant profile pointer in the device after
adjusting the reference count.