Bug 699676 "PDF interpreter can leave dangerous operators available"
The problem here is that certain PDF functions require the ability to
use some of the non-standard, and dangerous, operators/procedures
from the PostScript interpreter.

If we leave those functions readable then a malicious PostScript program
could instantiate the PDF interpreter, inspect the packedarrray and
copy the otherwise unobtainable operator/function, then use it for
mischief.

By making the PDF functions executeonly its impossible to read the
function contents, which prevents this kind of abuse.