Added Alternates support, tweaked handling of streams
The only file I found to test this with is:
   tests_private/pdf/PDF_1.7_FTS/fts_17_1713.pdf

In that file, there are two images, and one of them doesn't render due
to a colorspace issue, but the other image (the color one) displays on
the right side now.

Also, tweaked the handling of streams.  It will SEEK the stream each
time it needs to start rendering an image or mask, so caller needs to
preserve the stream offset if they care.  inline_image should be okay
because it doesn't support /Alternates or /Masks (which are the things
that have to mess with the stream position).