PDF interpreter - remove a warning message
Commit 	f6bcc0dfc173b80fddab4fbf10afb34332ff1112 added extra checking
and processing to deal with the (illegal) case of a Text block which
includes a 'q' operator.

As part of that commit, we added a warning message so that users would
be able to tell there was a problem. Unfortunately, there are cases
which are perfectly legal (eg running a Pattern PaintProc to fill the
text) where a q can apparently occur inside a text object.

We can't differentiate between executing a 'q' in such cases, so the
only option is to remove the warning. This does not affect the actual
operation of the code, it still copes with a 'q' inside text objects,
whether legally or not.