Bug 699720: Change available buffer space to int from uint
sbufavailable() returns the number of bytes still available in the stream
buffer. This can end up as negative at EOF (after we've read the final byte in
the buffer, we read ptr is moved to the next byte, past the end of the buffer -
that should never be read, however).

We can safely use a signed integer since this is buffer space, *not* the total
bytes available from the stream