Move from systemdict flag to internal operator flag for UEL handling.
Rather than relying on "PS_INTERP_ACT_ON_UEL" being in the
systemdict, put a flag in the gs_lib_ctx.

This flag is 0 by default, and we provide a gsapi_act_on_uel()
call to set it to 1.

We have a new .actonuel operator to get the value,
and we modify .forceinterp_exit and the PS setup code to use
this instead.