Bug 699740: Better handle failing to set the nulldevice
When pushing the nulldevice, if gs_setdevice_no_erase() failed, that *could*
mean we end up freeing the device referenced in the graphics state.

If that happens, search back through the saved gstates to find the initial
graphics state, and pull the initial device from there.

This allows us to return a Fatal error code, and exit cleanly.