Bug 699753: Windows uninstall: remove all installed files etc
This was partly caused by a subtlety of the nsis installer syntax which means
the '/r' option for the 'File' operation will match and install all
files/directories in the file heirarchy that match the search term, unless they
are explicitly excluded.

For example:
File /r doc

will end up matching:
doc/
something/doc
another/thing/doc

So, ensure we exclude subtrees so we no longer accidentally include files we
don't intend.

Also, add the missing removals, so we do remove eveything we intended to
install.

Lastly add in (and include removal of!) the Resource/* tree, and the iccprofiles
directories - both of which users have asked to be included.