pdfwrite - support fill+stroke primitive
Says it all really. Add a new device method to handle fill+stroke as
a primitive. If for any reason (eg reverting to old versions of PDF)
we can't, or may not be able to, emit the primitive, fall back to
handling it as a fill followed by a stroke.