Change bbox computation in pdf14_fill_stroke_path
This uses the pgs->clip_path, or the device width/height intersected
with the pcpath->outer_box (if non NULL), intersected with the ppath
expanded bbox. Note that pgs->path bbox is NOT used, since that is not
used in the calls to pdf14_fill_path and pdf14_stroke_path.