Bug 699777 " -dFILTERIMAGE often loses text"
The pdfwrite device's implementation of transparency relies upon the
image for a Softmask being run, and pushing a clip, which increments the
viewer state. However if we have -dFILTERIMAGE in place, then we won't
do this (its in the pdfwrite implementation of begin_typed_image).

When we close the transparency mask, this then leads to the viewer state
stack being out by one.

There's no easy way to address this, so when we go to restore the viewer
state, if it would fall off the bottom then check to see if FILTERIMAGE
is true. If it is, then just don't return an error.

This is a rare condition, it relies upon the transparency being nested
quite deeply in groups (in this case the SoftMask is nested about four
layers deep in spurious transparency).