Fix SEGV in gs_swapcolors used during clist playback (pdf14_fill_stroke_path).
The gs_swapcolors cannot be used during clist playback since the pgs doesn't
have enough information populated. The pdf14_fill_path and pdf14_stroke_path
use the pgs->overprint to set the overprint value in the pdf14_device for
the marking operation (see pdf14_set_marking_params).

In order to set the pgs->overprint for the pdf14_stroke_path call, we
need pgs->stroke_overprint, so this has to be added to the clist set of
gstate parameters that are passed through the clist.

Add this to the "misc2" set of parameters (this required adding another
byte since all the bits were used).