Change "executeonly" to throw typecheck on gstatetype and devicetype objects
PS doesn't really have a "devicetype", but it is not listed as an
acceptable argument type to "executeonly", so we probably should not
accept it.

"gstate executeonly" throws typecheck in Acrobat, so we should do the same.

Tested the other "access related" functions (noaccess, readonly,
rcheck, wcheck, xcheck) and all of these accept "gstate" as an
argument, even though the PLRM says they should not.  Also our initialization
code does things with wcheck and readonly on devices and gstates.

So, this fix is only for "executeonly".