Update language recognition mechanism.
Rather than each language implementation offering a single string
to be matched, they now offer a function to match the given
string.

This allows us to have the PJL recognising routine accept a
single <CR><LF> at the top of a PJL block.

Technically, this is against the spec, which very clearly says
that any UEL must be followed instantly by "@PJL" rather than
there being any whitespace there. Sadly, the file from bug 693269
sends UEL<CR><LF>@PJL... and so outputs a page of PJL in the
middle of the job without this fix.

Tests on Henry's HP printer show that his printer at least matches
this behaviour.

It is hard to see that there can be that many real world files out
there that will be adversely affected by this fix.

This means we now behave the same on bug 693269 as we did prior to
making our language switching actually switching languages.