Bug #699801 "zparse_dsc_comments can crash with invalid dsc_state"
Although zparse_dsc_comments() does check the types of its operands, it
wasn't checking the return value from dict_find_string properly. It was
effectively assuming that the dictionary parameter was the *correct*
dictionary and would contain the key/value pair it needed.

Here we check to see if the key has not been found and throw an error
if so.