Catch errors and cleanup stack on statusdict page size definitions
This was also found with the sequence from Bug 699794, if .setpagesize
fails, we need to clean up the two integers (page width, height) then
signal the error. Since some definitions "load" another (e.g. b5 loads
isob5) the command name in the error won't be correct for those, but
it will be close. If anybody cares, we could fix it.