device subclassing - fix use of put_image in memory devices
Bug 699799 "gswin64c.exe has stopped working"

The cause of this is truly ugly. It seems that the put_image device
method, uniquely, expects that the device parameter it is passed is not
the device for which it was intended, but is instead a 'memory'
device.

All the implementations of put_image start by casting the gx_device *
to a gx_device_memory *

This is problematic, because it completely disrupts any attempt at a
device chain. There's no way to have a device's method check to see if
it has a child device, if its not given its owe device to work with.

This commit hacks around the problem by descending to the terminating
device, and copying the put_image call from there into the memory device
which is what the pdf14 compositor will then use.

This needs fixing properly.